requirements.txt
Flask
google-generativeai
dotenv
Flask-SQLAlchemy
Flask-Login
Flask-Mail
itsdangerous
SQLAlchemy